#53 Modified api/AgentAPI.java, waiting for update at the admin side. Changed agentId from int to Long.

+++ helpdesk-data/src/main/java/hu/schonherz/training/helpdesk/data/entity/MessageEntity.java
@@ -18,7 +18,7 @@ public class MessageEntity extends BaseEntity {
 
     private static final long serialVersionUID = 1L;
 
-    private int agentId;
+    private Long agentId;
 
     private String clientId;
